chore(migrator): refactor LEGACY_ID_VAR_NAME qa -> core related to: camunda/camunda-bpm-platform/issues/5299
